refactor: Move error pages to `app/view/Error/Pages`
