+++ src/app/view/Error/ErrorScreen.tsx
@@ -4,11 +4,11 @@ import { Linking } from 'react-native'
 import { changeBarColors } from 'react-native-immersive-bars'
 import { WebViewMessageEvent } from 'react-native-webview/lib/WebViewTypes'
 
-import { CozyBlockedPage } from '/components/webviews/CozyBlockedPage'
-import { CozyNotFoundPage } from '/components/webviews/CozyNotFoundPage'
-import { CozyNotOnboardedPage } from '/components/webviews/CozyNotOnboardedPage'
-import { OfflinePage } from '/components/webviews/OfflinePage'
-import { ManagerErrorPage } from '/components/webviews/ManagerErrorPage'
+import { CozyBlockedPage } from '/app/view/Error/Pages/CozyBlockedPage'
+import { CozyNotFoundPage } from '/app/view/Error/Pages/CozyNotFoundPage'
+import { CozyNotOnboardedPage } from '/app/view/Error/Pages/CozyNotOnboardedPage'
+import { ManagerErrorPage } from '/app/view/Error/Pages/ManagerErrorPage'
+import { OfflinePage } from '/app/view/Error/Pages/OfflinePage'
 import { SupervisedWebView } from '/components/webviews/SupervisedWebView'
 import { routes } from '/constants/routes'
 import { goBack, reset } from '/libs/RootNavigation'
