+++ src/app/view/Error/Pages/CozyNotOnboardedPage.js
