+++ src/app/view/Error/Pages/ManagerErrorPage.js
