+++ src/app/view/Error/Pages/OfflinePage.js
